training/pyproject.toml
[project]
name = "fractal-gan-training"
version = "0.1.0"
description = "FastGAN training for fractal flame generation"
requires-python = ">=3.11"
dependencies = [
    "torch>=2.4",
    "torchvision>=0.19",
    "einops>=0.8.0",
    "pillow>=10.0",
    "numpy>=1.26",
    "scipy>=1.11",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "click>=8.1",
    "tensorboard>=2.15",
    "clean-fid>=0.1.35",
]

[tool.setuptools.packages.find]
include = ["model*", "data*", "utils*"]

[project.scripts]
train-gan = "train:main"
